+++ client/src/components/Header.tsx
@@ -11,12 +11,12 @@ export default function Header() {
   ];
 
   return (
-    <header className='sticky top-0 z-50'>
+    <header className='sticky top-0 z-100'>
       <nav className='bg-teal-800 border-b border-gray-200 shadow-sm w-full pt-2 relative'>
         {/* Absolute -  Login Button */}
         <a
           href='/login'
-          className='absolute right-8 top-8 transform px-4 h-10 rounded-3xl bg-slate-300 flex items-center justify-left cursor-pointer hover:bg-slate-100 hover:text-teal-800 transition duration-500 active:scale-95'
+          className='absolute right-8 top-8 transform px-4 h-8 rounded-3xl bg-slate-300 flex items-center justify-left cursor-pointer hover:bg-slate-100 hover:text-teal-800 transition duration-500 active:scale-95'
         >
           <img
             src={logo_login}
@@ -34,21 +34,21 @@ export default function Header() {
             <img
               src={logo_atrium}
               alt='Atrium'
-              width='64'
-              height='64'
+              width='48'
+              height='48'
             />
-            <h1 className='text-4xl font-bold text-slate-200 font-serif px-2 hover:text-white transition duration-900'>
+            <h1 className='text-2xl font-bold text-slate-200 font-serif px-2 hover:text-white transition duration-900'>
               Atrium
             </h1>
           </a>
         </div>
         {/* Static - Nav Items Container */}
         <div className='flex items-center justify-center'>
-          <ul className='flex justify-center h-12'>
+          <ul className='flex justify-center h-8'>
             {list_nav_items.map((item: TNavItem) => (
               <li
                 key={item.label}
-                className='flex items-center px-8 py-4 text-xl text-slate-300 font-serif font-bold cursor-pointer hover:border-b-6 hover:border-teal-600 transition duration-900'
+                className='flex items-center px-8 py-2 text-l text-slate-300 font-serif font-bold cursor-pointer hover:border-b-6 hover:border-teal-600 transition duration-900'
               >
                 <a
                   href={item.href}